+++ src/main/java/cn/org/shelly/edu/model/resp/ProposalRoundTeamScoreResp.java
@@ -2,10 +2,12 @@ package cn.org.shelly.edu.model.resp;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProposalRoundTeamScoreResp {
     private Long teamId;
     private String leaderName;  // 小组组长姓名
-    private Integer score;      // 积分
+    private BigDecimal score;      // 积分
     private Integer rank;       // 排名
 }
